Moves to Avatar component from LunaUI
