--- app/components/ui/avatar.scss
@@ -1,46 +1,0 @@
-@use '../../styles/pxToRem.scss' as *;
-
-.avatar-container {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  color: var(--font-color);
-
-  &:hover {
-    cursor: pointer;
-    color: var(--primary-500);
-  }
-
-  .avatar-default:hover {
-    color: var(--font-color);
-  }
-}
-
-.avatar {
-  width: pxToRem(40);
-  height: pxToRem(40);
-  border-radius: var(--radius-pill);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  margin-right: pxToRem(6);
-}
-
-.avatar-default {
-  width: pxToRem(40);
-  height: pxToRem(40);
-  border-radius: var(--radius-pill);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  margin-right: pxToRem(6);
-  background-color: var(--primary-700);
-  font-size: var(--font-xl);
-  color: var(--font-color);
-}
-
-.avatar-username {
-  font-size: var(--font-xl);
-  font-weight: bold;
-  transition: var(--transition-base);
-}
